Adding new IPTV app for Linux
New IPTV desktop app for linux made with QT C++. Didn't see any CONTRIBUTION.md file in this repo. So, If this is not correct way to do so, you may correct me.
I am the author of this project. Binary file may not work on any computer which does not have QT installed so they may compile it. (I really don't know how to deploy a QT App, Sorry)